Esp0

+++ fse-trabalho3-adnecaiogabriel/include/dht11.h
@@ -10,9 +10,9 @@ enum dht11_status {
 };
 
 struct dht11_reading {
-    int status;
-    int temperature;
-    int humidity;
+    float status;
+    float temperature;
+    float humidity;
 };
 
 void DHT11_init(gpio_num_t);
